Option for drawing Surround Background
Adding a feature that lets you subtract background on a spine-by-spine basis by drawing a larger, concentric circle around the target ROI. This should help in cases where invading neurites etc. that pass over the ROI interfere with the signal. However, this probably also interferes with the dendrite subtraction method, as this would sometimes subtract out a portion of the dendrite, if the spine is close enough to the shaft.